refactor(physics): delete unused super folder /system.
