--- src/systems/mod.rs
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: MIT
-// Copyright (c) 2025 Tingxu Chen
-// Author: Tingxu Chen <tic128@pitt.edu>
-// Description: <Systems module and plugins>
-pub mod physics;
